python/gitwip.requirements.txt
gitdb==4.0.10
GitPython==3.1.37
markdown-it-py==3.0.0
mdurl==0.1.2
Pygments==2.16.1
rich==13.6.0
smmap==5.0.1
